advance date in later threads
The book had assumed the start date (when the book was compiled) was the
date for all threads.  This won't be accurate unless the group plays
through the entire chronicle in a fortnight.

Now the later threads advance the date.  The dates will never be
accurate, but the little bits which rely on Fenestra's dates (monster's
stats, available food, descriptions of wind) will probably match a
little closer to a real game.